refactor: 모델 파일을 각 모듈별 디렉터리로 이동
