+++ lib/app/api_client.dart
@@ -3,15 +3,15 @@ import 'dart:convert';
 import 'dart:developer';
 import 'dart:io';
 
-import 'package:duty_it/app/models/alarm_settings.dart';
-import 'package:duty_it/app/models/app_notification.dart';
-import 'package:duty_it/app/models/app_user.dart';
-import 'package:duty_it/app/models/event.dart';
-import 'package:duty_it/app/models/event_type.dart';
-import 'package:duty_it/app/models/host.dart';
-import 'package:duty_it/app/models/login_result.dart';
-import 'package:duty_it/app/models/server_fail.dart';
-import 'package:duty_it/app/models/sort_direction.dart';
+import 'package:duty_it/app/modules/settings/models/alarm_settings.dart';
+import 'package:duty_it/app/modules/notifications/models/app_notification.dart';
+import 'package:duty_it/app/core/models/app_user.dart';
+import 'package:duty_it/app/core/models/event.dart';
+import 'package:duty_it/app/core/enums/event_type.dart';
+import 'package:duty_it/app/core/models/host.dart';
+import 'package:duty_it/app/modules/login/models/login_result.dart';
+import 'package:duty_it/app/core/models/server_fail.dart';
+import 'package:duty_it/app/core/enums/sort_direction.dart';
 import 'package:duty_it/app/services/auth/auth_service.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
