+++ lib/app/core/enums/event_type.dart
