+++ lib/app/core/models/app_user.freezed.dart
