+++ lib/app/core/models/event.freezed.dart
