+++ lib/app/core/models/host.dart
