+++ lib/app/core/models/host.freezed.dart
