+++ lib/app/core/models/server_fail.dart
