+++ lib/app/core/models/server_fail.freezed.dart
