+++ lib/app/core/utils/events/event_bookmark_event.dart
@@ -1,0 +1,3 @@
+import 'package:duty_it/app/core/utils/events/app_event.dart';
+
+class EventBookmarkEvent extends AppEvent {}
