+++ lib/app/modules/login/models/login_result.freezed.dart
