+++ lib/app/modules/notifications/models/app_notification.freezed.dart
