+++ lib/app/services/auth/auth_service.dart
@@ -2,7 +2,7 @@ import 'dart:developer';
 
 import 'package:duty_it/app/api_client.dart';
 import 'package:duty_it/app/core/utils/app_utils.dart';
-import 'package:duty_it/app/models/app_user.dart';
+import 'package:duty_it/app/core/models/app_user.dart';
 import 'package:duty_it/app/modules/calendar/controllers/calendar_view_controller.dart';
 import 'package:duty_it/app/services/auth/models/social_login_result.dart';
 import 'package:duty_it/app/services/auth/strategies/apple_login_strategy.dart';
